+++ Sensors@claudiux/files/Sensors@claudiux/scripts/get_disk_temp.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/usr/bin/env sh
 # Author: @claudiux
 # Contributor (NVMe devices): @kriegcc