+++ Sensors@claudiux/files/Sensors@claudiux/scripts/get_disk_temp_ORIG.sh
@@ -1,3 +1,3 @@
-#!/bin/sh
+#!/usr/bin/env sh
 echo -n $(sudo smartctl -A /dev/$1 | grep 'Temperature_Cel' | awk '{print $10}')
 exit 0