+++ multicore-sys-monitor@ccadeptic23/files/multicore-sys-monitor@ccadeptic23/scripts/get-cpu-data.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 echo -n $(cat /proc/cpuinfo | grep -E "MHz|bogomips")