+++ sound150@claudiux/files/sound150@claudiux/scripts/rm_tmp_files.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 ### Removes cover art files.
 
 rm -f /tmp/*.mediaplayer-cover