+++ brightness-and-gamma-applet@cardsurf/files/brightness-and-gamma-applet@cardsurf/CHANGELOG.md
@@ -1,3 +1,17 @@
+### v2.1.3~20250810
+* Fixes [#7510](https://github.com/linuxmint/cinnamon-spices-applets/issues/7510)
+* Return of option in General tab: "Apply when the number of monitors changes".
+
+### v2.1.2~20250807
+* Fixes [#7452](https://github.com/linuxmint/cinnamon-spices-applets/issues/7452)
+
+### v2.1.1~20250806
+* Fixes [#7452](https://github.com/linuxmint/cinnamon-spices-applets/issues/7452)
+
+### v2.1.0~20250805
+* Fixes [#7452](https://github.com/linuxmint/cinnamon-spices-applets/issues/7452)
+* New option in General tab: "Apply when the number of monitors changes".
+
 ### v2.0.1~20250714
 * Tooltip bugfix.